pyproject.toml
[build-system]
requires = ['flit_core >=3.4,<4']
build-backend = 'flit_core.buildapi'

[project]
name = 'aiida-quantumespresso-ph'
dynamic = ['description', 'version']
authors = [
    {name = 'Sebastiaan P. Huber', email = 'mail@sphuber.net'},
    {name = 'Marnik Bercx', email = 'marnik.bercx@epfl.ch'},
    {name = 'Michele Kotiuga', email = 'michele.kotiuga@epfl.ch'}
]
readme = 'README.md'
license = {file = 'LICENSE.txt'}
classifiers = [
    'Development Status :: 3 - Alpha',
    'Framework :: AiiDA',
    'License :: OSI Approved :: MIT License',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
]
keywords = ['aiida', 'workflows']
requires-python = '>=3.9'
dependencies = [
    'aiida-quantumespresso>=4.5',
]

[project.urls]
Source = 'https://github.com/aiidateam/aiida-quantumespresso-ph'

[project.optional-dependencies]
pre-commit = [
    'pre-commit~=2.2',
    'pylint==2.13.7',
    'pylint-aiida~=0.1',
]
tests = [
    'pgtest~=1.3',
    'pytest>=6.0',
    'pytest-regressions~=2.3'
]

[project.entry-points.'aiida.calculations']
'quantumespresso_ph.distribute_qpoints' = 'aiida_quantumespresso_ph.calculations.functions.distribute_qpoints:distribute_qpoints'
'quantumespresso_ph.recollect_qpoints' = 'aiida_quantumespresso_ph.calculations.functions.recollect_qpoints:recollect_qpoints'
'quantumespresso_ph.merge_para_ph_outputs' = 'aiida_quantumespresso_ph.calculations.functions.merge_para_ph_outputs:merge_para_ph_outputs'

[project.entry-points.'aiida.workflows']
'quantumespresso.dynamical_matrix' = 'aiida_quantumespresso_ph.workflows.dynamical_matrix:DynamicalMatrixWorkChain'
'quantumespresso.ph_interpolate' = 'aiida_quantumespresso_ph.workflows.ph_interpolate:PhInterpolateWorkChain'
'quantumespresso_ph.ph.main' = 'aiida_quantumespresso_ph.workflows.ph.main:PhWorkChain'
'quantumespresso_ph.ph.parallelize_qpoints' = 'aiida_quantumespresso_ph.workflows.ph.parallelize_qpoints:PhParallelizeQpointsWorkChain'

[tool.flit.module]
name = 'aiida_quantumespresso_ph'

[tool.flit.sdist]
exclude = [
    '.github/',
    'examples/',
    'tests/',
    '.gitignore',
    '.pre-commit-config.yaml',
]

[tool.flynt]
line-length = 120
fail-on-change = true

[tool.isort]
force_sort_within_sections = true
include_trailing_comma = true
line_length = 120
multi_line_output = 3

[tool.pydocstyle]
ignore = [
    'D104',
    'D202',
    'D203',
    'D213'
]

[tool.pylint.basic]
good-names = [
    'i',
    'j',
    'k',
    'SsspFamily',
    'StructureData',
    'SinglefileData',
    'UpfData',
    'KpointsData',
    'StructureData',
    'HpCalculation',
    'PwCalculation',
    'PhCalculation',
]

[tool.pylint.master]
load-plugins = ['pylint_aiida']

[tool.pylint.format]
max-line-length = 120

[tool.pylint.messages_control]
disable = [
    'bad-continuation',
    'duplicate-code',
    'import-outside-toplevel',
    'inconsistent-return-statements',
    'too-many-arguments',
]

[tool.pytest.ini_options]
minversion = '6.0'
testpaths = [
    'tests',
]

[tool.yapf]
align_closing_bracket_with_visual_indent = true
based_on_style = 'google'
coalesce_brackets = true
column_limit = 120
dedent_closing_brackets = true
indent_dictionary_value = false
split_arguments_when_comma_terminated = true
